Extract merged-slot compiler fixtures
Move the merged-slot naming samples beside their sole owning test while preserving the compiler-produced source shape.
