+++ src/ILInspector.Decompiler.Tests/CfgSampleClass.cs
@@ -5952,27 +5952,6 @@ public static class EnumCastSamples
         => new[] { (System.StringComparison)4, System.StringComparison.OrdinalIgnoreCase };
 }
 
-// Issue #1767: a value produced at a stack slot (the object-merged `cond ? null
-// : value` ternary) and consumed at a control-flow merge where the slot is typed
-// `string` (a field store) must share ONE local name. Keying slot names on
-// (slot, type) split them into `object S_1` (store) and `string S_1_1` (load),
-// so the consumer read an unassigned local (CS0165) and the value was dropped.
-public static class MergedSlotStrU
-{
-    public static bool IsNullOrEmpty(string s) => s == null || s.Length == 0;
-}
-
-public class MergedSlotProbe
-{
-    private string? _fmt;
-
-    public string? MergedFormat
-    {
-        get => _fmt;
-        set => _fmt = MergedSlotStrU.IsNullOrEmpty(value!) ? null : value;
-    }
-}
-
 // Issue: a user-defined operator with `in`/`ref` parameters is called with the
 // operands' addresses (ldarga/ldloca). The operator must spell as `a != b`, not
 // `(ref a) != (ref b)` — the latter is CS1525 "Invalid expression term 'ref'".
